refactor: extract PrivacyDataSessionStream.swift from PrivacyDataExportTypes.swift
Splits the lazy session-stream type out of PrivacyDataExportTypes.swift
so that file holds only Codable/Encodable snapshot types. The stream
is a lazy iterator wrapper (see #508 rationale in the comment).

Byte-identical move of lines 59-81. Both files import Foundation.

PrivacyDataExportTypes.swift: 81 → 57 lines. Tests: 667.
